run FE and BE tests
